LEN_TRIM inline implementation + make character length index type
- Use fir::iter_while operation to provide a scalar implementation
of LEN_TRIM that is inlined. It is just a loop that goes from
the right most character to the first one and stop as soon as
a character is not a blank.
- Change getLengthType to return the indexType. It makes loop
manipulation of character easier. A few functions that expected
i64 for the length are modified to support whatever integer type
getLengthType is returning.

Conflicts:
	flang/lib/Lower/FIRBuilder.cpp

Conflicts:
	flang/LAPACK-bugs.txt