[refactor] thread the location information properly from the converter through all of the builder.
The current location is maintain as a SPOT in the converter. The helper classes are lightweight and ephemeral and often maintain a copy. The FirOpBuilder is long-lived (for the entirety of building a FuncOp) and should not maintain a on-the-side state of the location.

These changes will fix all the bugs where location information was coming out incorrectly because of bad cached state.