Fix call site mangling of external procedures
This addresses issue 138 of f18-llvm-project:
- Symbol with ProcEntityDetails are always external procedures unless they are
  procedure pointers/dummy arguments. Add a check for pointers dummy and remove
  the IsExternal check that is only answering true if the procedure appeared in
  an explicit EXTERNAL statement.

On top of issue 138:

- Another issue was that when an external program is defined
inside an interface, then the related Symbol has SubprogramDetails, so
isExternal must be checked in this case. See bar3 in added tests for an
example of what was failing.

- Looking at ProcEntityDetails more closely, I understand that the
related interface symbol has nothing to do with the actual definition,
so remove this from the logic looking for the callee definition
source location.