Pre-declare FuncOp of all procedures declared in a program unit before lowering bodies
The need to do this arise from the implicit interface rules that lead to situation where
the signature of a a function on call site is different from the definition signature.
This is particularly important when the first usage of a function is to be passed as
dummy argument, in which case we do know how many arguments the FuncOp must have,
leading to crashes if we later try to define it with one or more arguments.

To ensure that the definition signature prevail if it is accessible, pre-declare all
mlir::FuncOp before lowering the bodies. This is done by making a pass on the top
level of the PFT before the pass to lower the bodies. The CallInterface is refactored a
bit to help do that.