Fix some problems on macos
`timeval.tv_usec` is int on macos, not long, so there was a warning about
using `snprintf` with %03ld format conversion. Use an explicit type of
std::maxint_t and %03jd to avoid warnings regardless of the type.

`OSType::Darwin` was an unknown target. Do the same thing for it as for Linux.